refactor: 프로필 공통구조만 두고 outlet 받아서 사용. 프로필과 프로필수정 컴포넌트 분리 수정 라우터 새로 만들고 페이지 이동시 스크롤바 위치 기억하던거 useRef, useEffect로 항상 0으로 받아올 수 있게 함.
