+++ src/components/mypage/activityLog/ActivityLog.tsx
@@ -2,9 +2,5 @@ import { ACTIVITY_FILTER } from '../../../constants/myPageFilter';
 import ContentTab from '../ContentTab';
 
 export default function ActivityLog() {
-  return (
-    <>
-      <ContentTab $justifyContent='space-around' filter={ACTIVITY_FILTER} />
-    </>
-  );
+  return <ContentTab $justifyContent='space-around' filter={ACTIVITY_FILTER} />;
 }
