+++ src/hooks/useMyInfo.ts
@@ -44,7 +44,9 @@ export const useEditMyProfileInfo = (
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: myInfoKey.myProfile });
       handleModalOpen(MODAL_MESSAGE.myProfileSuccess);
-      navigate(ROUTES.mypage);
+      setTimeout(() => {
+        navigate(ROUTES.mypage);
+      }, 1500);
     },
     onError: () => {
       handleModalOpen(MODAL_MESSAGE.myProfileFail);
